Update utopia_ascension_l_english.yml
a brief gateway into the Shroud.\n\nNow that we have established an initial psionic link, we will be able to access the Shroud through the §HContacts§! menu in the future. To make any kind of meaningful forays into this realm and meet the beings who dwell there, we will likely have to make use of special psionic emitters to extend our psychic reach. This will consume a tremendous amount of energy

翻译不好